refactor(archtest): 迁移 webhook 规则族 → module-path-agnostic Check* + 缩 baseline [#1636]
MOVE webhook_hmac_funnel_test.go（WEBHOOK-HMAC-FUNNEL-01，子规则 A1 hmac.New allowlist / A2 常量时间比较 /
A3 sealed Signer-Verifier marker / B6 no raw secret slog）的检测逻辑到非 test 的 webhook_hmac_funnel.go
（CheckWebhookHMACFunnel + scanner helpers）；平台符号路径经 PlatformModulePath 派生
（line 209/294 bare github.com/ghbvf/gocell/kernel/webhook → webhookPkgPath）。

_test.go 瘦身为 dogfood Report() + 保留 _ReverseFixture（A1/A2 fire on testdata/webhook_hmac_violate/）+
_NoRawSecretSlog（B6 自检）。A3 由原 assert.True 正向断言等价转为 checkWebhookSealedMarkers 产 Diagnostic
（缺接口 / 缺 sealed marker → diag，Report 失败语义一致）。gocognit ≤15 经 collectSecretSlogArgs /
scanWebhookPkg / checkWebhookSealedMarkers 行为保持 helper 提取达成，无 //nolint。register=no
（gocell-internal-layout，扫 kernel/webhook，外部 vacuous）。行为零变更（verbatim MOVE + 恒等派生），
TypedOpts 无 Tags 原样保留。baseline 删 1 行（messaging 族 4 行全部 ratchet 完成）。

ref: golang.org/x/tools/go/analysis Analyzer
