+++ tools/archtest/testdata/module_path_funnel.baseline
@@ -65,5 +65,4 @@ tools/archtest/sessionrefresh_no_session_create_test.go
 tools/archtest/svctoken_caller_cell_test.go
 tools/archtest/test_polling_external_reason_literal_test.go
 tools/archtest/user_repo_conformance_enrollment_test.go
-tools/archtest/webhook_hmac_funnel_test.go
 tools/archtest/wrapper_location_test.go
